pyproject.toml
[project]
name = "digital-excellence-talent-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "flask==3.0.3",
    "flask-cors==4.0.0",
    "google-genai==0.3.0",
    "gunicorn==21.2.0",
    "supabase>=2.28.3",
]
